requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
WTForms==3.0.1
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
weasyprint==60.1
fonttools==4.42.1
html2canvas==1.4.1
qrcode==7.4.2
Pillow-SIMD==9.0.0.post1
python-magic==0.4.27
